Don't fix /dev/null permissions if it fails in creation
By doing so we'll be masking a symptom of the problem instead of fixing
it. If permissions of /dev/null fail, most likely other problems will
also happen (e.g. pts permissions). Keep /dev/null alarm (which leads
to X server starting problems) to alert us if something is wrong.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>